requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cleanup==6.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==6.0.0
phonenumbers==8.12.42
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
sqlparse==0.4.2
tzdata==2021.5
